Metadata attributes constant extraction
